+++ example/lib/main.dart
@@ -23,7 +23,7 @@ void main() {
   /// to be defined (or re-defined) can be done.
   Segment.setContext({
     'device': {
-      'token': 'testing'
+      'token': 'testing',
     }
   });
 
@@ -44,7 +44,6 @@ class MyApp extends StatelessWidget {
         body: Column(
           children: <Widget>[
             Spacer(),
-
             Center(
               child: FlatButton(
                 child: Text('TRACK ACTION WITH SEGMENT'),
@@ -60,31 +59,24 @@ class MyApp extends StatelessWidget {
                 },
               ),
             ),
-
             Spacer(),
-
             Center(
               child: FlatButton(
                 child: Text('Update Context'),
                 onPressed: () {
-                  Segment.setContext({
-                    'custom': 123
-                  });
+                  Segment.setContext({'custom': 123});
                 },
               ),
             ),
-
             Spacer(),
-
             Center(
               child: FlatButton(
                 child: Text('Clear Context'),
                 onPressed: () {
-                  Segment.setContext({ });
+                  Segment.setContext({});
                 },
               ),
             ),
-
             Spacer(),
 
             Center(